apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.116.1",
    "pipecat-ai[silero,websocket]>=0.0.31",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
